requirements.txt
Pillow==9.5.0
uuid==1.30
webptools==0.0.9
